agent/requirements.txt
langchain>=0.3.0
langchain-classic>=1.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
web3>=6.11.0
pydantic>=2.5.0
openai>=1.40.0